add tests for each banner type
